+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Partial documentation for hipFFT is available at [hipFFT].
 ### Added
 - Added File/Folder Reorg Changes with backward compatibility support using ROCM-CMAKE wrapper functions.
 - Packages for test and benchmark executables on all supported OSes using CPack.
+- Implemented hipfftMakePlanMany64 and hipfftGetSizeMany64.
 
 ## hipFFT 1.0.7 for ROCm 5.1.0